+++ core-codemods/src/test/java/io/codemodder/codemods/integration/baseimage/Dockerfile
